requirements.txt
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
filelock==3.13.1
fsspec==2024.2.0
geopandas==1.0.1
idna==3.10
inquirerpy==0.3.4
Jinja2==3.1.3
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pfzy==0.3.4
pillow==10.4.0
planetary-computer==1.0.0
prompt_toolkit==3.0.47
pydantic==2.9.2
pydantic_core==2.23.4
pyogrio==0.9.0
pyproj==3.6.1
pystac==1.10.1
pystac-client==0.8.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rpds-py==0.20.0
shapely==2.0.6
six==1.16.0
sympy==1.12
toml==0.10.2
torch==2.4.1+cpu
torchvision==0.19.1+cpu
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
